+++ ChangeLog.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - `e3nn.reduced_symmetric_tensor_product_basis(irreps: Irreps, order: int)`
+- `e3nn.IrrepsArray.filtered(keep: List[Irrep])`
 
 ## [0.9.0] - 2022-09-04
 ### Added